implement file lock

+++ server/app.py
@@ -1,6 +1,19 @@
+from pickle import FALSE
+from unittest import result
 from flask import Flask, jsonify, request
 from functions import *
 app = Flask(__name__)
+
+lockedFiles = []
+
+
+def isFileLocked(fileName):
+    found = False
+    for lockedFile in lockedFiles:
+        if lockedFile == fileName:
+            found = True
+            break
+    return found
 
 
 @app.post("/add_file")
@@ -12,8 +25,12 @@ def add_file():
 @app.delete("/del_file")
 def del_file():
     fileName = request.args.get('fileName', '')
-    print(fileName)
-    return jsonify(deleteFile(fileName))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(deleteFile(fileName))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.post("/add_dir")
@@ -25,13 +42,23 @@ def add_dir():
 @app.delete("/del_dir")
 def del_dir():
     dirName = request.args.get('dirName', '')
-    return jsonify(deleteDirectory(dirName))
+    if isFileLocked(dirName):
+        return jsonify({"error": "Directory already in use"})
+    lockedFiles.append(dirName)
+    result = jsonify(deleteFile(dirName))
+    lockedFiles.remove(dirName)
+    return result
 
 
 @app.get("/open_file")
 def open_file():
     fileName = request.args.get('fileName', '')
-    return jsonify(openFile(fileName))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(deleteFile(fileName))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.patch("/move_file_content")
@@ -40,7 +67,12 @@ def move_file_content():
     start = request.args.get("start", "")
     size = request.args.get("size", "")
     target = request.args.get("target", "")
-    return jsonify(moveContentWithinFile(fileName, start, size, target))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(moveContentWithinFile(fileName, start, size, target))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.get("/read_file")
@@ -48,7 +80,12 @@ def read_file():
     fileName = request.args.get('fileName', '')
     pos = request.args.get("pos", "")
     size = request.args.get("size", "")
-    return jsonify(readFromFile(fileName, pos, size))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(readFromFile(fileName, pos, size))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.patch("/write_file")
@@ -57,21 +94,32 @@ def write_file():
     pos = request.args.get("pos", "")
     text = request.args.get("text", "")
     appMode = request.args.get("appMode", "")
-    return jsonify(writeToFile(fileName, pos, text, appMode))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(writeToFile(fileName, pos, text, appMode))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.patch("/move_file")
 def move_file():
     fileName = request.args.get('fileName', '')
     newDir = request.args.get('newDir', '')
-    print(fileName, newDir)
-    return jsonify(moveFile(fileName, newDir))
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
+    lockedFiles.append(fileName)
+    result = jsonify(moveFile(fileName, newDir))
+    lockedFiles.remove(fileName)
+    return result
 
 
 @app.patch("/truncate_file")
 def truncate_file():
     fileName = request.args.get('fileName', '')
     size = request.args.get("size", "")
+    if isFileLocked(fileName):
+        return jsonify({"error": "File already in use"})
     return jsonify(truncateFile(fileName, size))
